Fixed issue 354.
The page view did not actually fail anymore, but I realized that it only 
did not, because subscriptions were not included in the list of adapters 
anymore. So I added subscriptions back and the problem reoccurred, which 
was due to the fact that subscription registrations do not have a name. 

Then I noticed another problem. Subscriptions often have `None` as its 
provided interface and not all code did expect that, so I fixed it. For 
some reason the code checking for `None` in the required adapters was 
already done.