First stab of making the static version of the apidoc actually work. It turned out that wget was not good enough for our case and did not rewrite all links that need to be rewritten.
So I developed a custom retrieval script. The advantage is that we can use 
the publisher directly, instead of brining up a Web server. Otherwise the 
script is a bit of a mess; no options, no documentation and lots of 
spaghetti code. I check it in in hope that people will help me improve it.

While I have not made a complete run yet, about 10500 links are correctly 
handled, which is about the first and second tier of links. This easily 
includes all links directly available from the menus.

Trying to generate the static version of the apidoc has surfaced several 
bugs, especially related to utility names. I will fix those during the 
freeze stage.