Try to fix the tests
Whoops: zope.interface and other modules no longer install their
documentation into the Python package directory -- they moved to
Sphinx-based docs that are present in the sdist but not included in
binary distributions.

zope.app.apidoc needs to be redesigned (to fetch the source packages
perhaps?  using mr.developer or something) or it needs to pin
zope.interface etc. to older versions (but then what use is it?).