Restore ++apidoc++ to the root.
In the tests, use the correct root folder, not the site folder.

But also, the traverser for the apidoc namespace forces the issue.
This prevents issues with shared ZODB connections.

Doing this allows us to correctly spider all the ++apidoc++ links
without running into circular parent chains or closed connection errors.