Fix various forms of namespace packages not finding children.
The 'zope' node in the tree was missing in pip installs.

Delegate installed/not-installed to the actual ZCML implementation so
that we don't wind up trying to run python 2/3 specific directives on
incorrect platforms, since that's more common now.

Also clean up dependencies and remove dep on zope.app.securitypolicy
which hasn't actually been ported to Python 3.